Added HLTB game name to scraped data
Finding the right game on HLTB when searching by game name is not as
good as I would like it to be, so I've decided to get the HLTB game name
as well. I plan on manually fixing some entries and keeping a database
of correct matches in the future, so this will help me find errors.