Better error handling & new achievementless games
Previously, if there was an error, the script would instantly quit out
and give just a vague error message. Now, it continues scraping and
reports the appid/name of the game causing a problem.

Currently, games with exclusively unicode characters can't be scraped
which will need to be fixed.

Will also need to restructure scrape_steam_data in steam_utils.