Restructured scrape_steam_data to handle one game
With the updates made in the last change, we no longer need to
loop through a list of games in scrape_steam_data since that is
done in main.

I have also stopped stripping unicode from titles in steam_utils since
there are many games that are strictly/mostly unicode characters.
I will instead worry about that sort of thing in hltb_utils. That is
where unicode tends to cause a problem.