Fixed a search to stop a list index error
I was getting a `list index out of range` error for unicode-named games.
This was happening because I was stripping out unicode to search HLTB,
and then the edition search was trying to split an empty string:

`if not results_list and 'game_name.split()[-1].lower() == "edition"':
        results_list = hltb().search(game_name.rsplit(' ', 2)[0])`

I changed it to check if the word `edition` was in the title to avoid
this.